projetweb/src/requirements.txt
pymysql
mysqlclient
mysql-connector-python

gunicorn
Flask
flask-sqlalchemy